+++ docs/pages/components/progress/examples/ExSimple.vue
@@ -1,0 +1,17 @@
+<template>
+    <div class="example-component">
+        <section>
+            <b-progress :value="value"></b-progress>
+        </section>
+    </div>
+</template>
+
+<script>
+    export default {
+        data() {
+            return {
+                value: 0.4
+            }
+        }
+    }
+</script>