+++ docs/router/index.js
@@ -41,6 +41,7 @@ const router = new Router({
                 route('documentation/tabs', 'components/tabs/Tabs'),
                 route('documentation/collapse', 'components/collapse/Collapse'),
                 route('documentation/panel', 'components/panel/Panel'),
+                route('documentation/progress', 'components/progress/Progress'),
                 route('documentation/tag', 'components/tag/Tag'),
                 route('documentation/dialog', 'components/dialog/Dialog'),
                 route('documentation/toast', 'components/toast/Toast'),