+++ src/components/progress/Progress.spec.js
@@ -1,0 +1,10 @@
+import { shallow } from '@vue/test-utils'
+import BProgress from '@components/panel/Progress'
+
+describe('BProgress', () => {
+    it('is called', () => {
+        const wrapper = shallow(BProgress)
+        expect(wrapper.name()).toBe('BProgress')
+        expect(wrapper.isVueInstance()).toBeTruthy()
+    })
+})