+++ src/components/progress/index.js
@@ -1,0 +1,13 @@
+import Progress from './Progress'
+
+import { use, registerComponent } from '../../utils/plugins'
+
+const Plugin = {
+    install(Vue) {
+        registerComponent(Vue, Progress)
+    }
+}
+
+use(Plugin)
+
+export default Plugin